update reserves endpoint with new payload & verification
