+++ packages/sources/bitgo-reserves/src/config/index.ts
@@ -6,11 +6,6 @@ export const config = new AdapterConfig({
     type: 'string',
     default: 'https://reserves.usdstandard-test.com/por.json',
   },
-  API_ENDPOINT2: {
-    description: 'An API endpoint for Data Provider',
-    type: 'string',
-    required: true,
-  },
   VERIFICATION_PUBKEY: {
     description: 'Public key used for verifying data signature',
     type: 'string',
