+++ packages/sources/bitgo-reserves/src/endpoint/index.ts
@@ -1,2 +1,1 @@
 export { endpoint as reserves } from './reserves'
-export { endpoint as reserves2 } from './reserves2'
