--- packages/sources/bitgo-reserves/src/endpoint/reserves2.ts
@@ -1,19 +1,0 @@
-import { AdapterEndpoint } from '@chainlink/external-adapter-framework/adapter'
-import { InputParameters } from '@chainlink/external-adapter-framework/validation'
-import { SingleNumberResultResponse } from '@chainlink/external-adapter-framework/util'
-import { config } from '../config'
-import { httpTransport } from '../transport/reserves2'
-
-export const inputParameters = new InputParameters({})
-
-export type BaseEndpointTypes = {
-  Parameters: typeof inputParameters.definition
-  Response: SingleNumberResultResponse
-  Settings: typeof config.settings
-}
-
-export const endpoint = new AdapterEndpoint({
-  name: 'reserves2',
-  transport: httpTransport,
-  inputParameters,
-})
