+++ packages/sources/bitgo-reserves/test/integration/fixtures.ts
@@ -8,7 +8,8 @@ export const mockResponseSuccess = (): nock.Scope =>
     .reply(
       200,
       () => ({
-        totalReserve: 1234567.89,
+        data: '{"reserveAmount":"12345678.90","cashReserve":"2345678.90","investedReserve":"10000000.00","lastUpdated":"2024-12-10T01:23:45Z"}',
+        dataSignature: 'testsig',
         lastUpdated: '2024-10-01T01:23:45Z',
       }),
       [
